requirements.txt
Django>=4.1.6,<4.2
django-environ>=0.9.0,<0.10
psycopg2>=2.9.5,<2.10
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.2.2,<5.3
social-auth-app-django>=5.0.0,<5.1
django-cors-headers>=3.13.0,<3.14
django-templated-mail>=1.1.1,<1.2
Pillow>=9.4.0,<9.5
